Rename `network-protocol` -> `net-protocol`.
