+++ src/builtin-applications/package.json
@@ -17,7 +17,7 @@
     "@this/fs-util": "*",
     "@this/loggy": "*",
     "@this/net-util": "*",
-    "@this/network-protocol": "*",
+    "@this/net-protocol": "*",
     "@this/typey": "*",
     "express": "^4.18.2"
   }
