+++ src/builtin-services/export/RequestLogger.js
@@ -8,7 +8,7 @@ import { BaseService } from '@this/sys-framework';
 import { Rotator } from '@this/sys-util';
 import { Moment } from '@this/data-values';
 import { IntfLogger } from '@this/loggy';
-import { IntfRequestLogger } from '@this/network-protocol';
+import { IntfRequestLogger } from '@this/net-protocol';
 
 
 /**
