+++ src/net-protocol/README.md
@@ -1,5 +1,5 @@
-@this/network-protocol
-======================
+@this/net-protocol
+==================
 
 This module handles the "protocol level" of the system. In this case, this means
 handling connections accepted from server sockets, passing through high-level
