+++ src/net-protocol/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@this/network-protocol",
+  "name": "@this/net-protocol",
   "version": "0.0.1",
   "type": "module",
   "private": true,
