+++ src/sys-framework/package.json
@@ -17,7 +17,7 @@
     "@this/collections": "*",
     "@this/loggy": "*",
     "@this/net-util": "*",
-    "@this/network-protocol": "*",
+    "@this/net-protocol": "*",
     "@this/typey": "*",
     "pem": "^1.14.8"
   }
